Step1: DetailView를 Controller와 분리
