+++ WeatherForecast/WeatherForecast/Source/Scene/WeatherDetail/WeatherDetailViewController.swift
@@ -6,10 +6,11 @@
 
 import UIKit
 
-class WeatherDetailViewController: UIViewController {
+final class WeatherDetailViewController: UIViewController {
     private let weatherForecastInfo: WeatherForecastInfo
     private let cityInfo: City
     private let tempUnit: TempUnit
+    private let contentView = WeatherDetailView()
     
     private let imageService: WeatherImageService
     
@@ -31,98 +32,46 @@ class WeatherDetailViewController: UIViewController {
     }
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         initialSetUp()
+    }
+    
+    override func loadView() {
+        view = contentView
     }
     
     private func initialSetUp() {
         view.backgroundColor = .white
+        contentView.spacingView.backgroundColor = .clear
+        contentView.spacingView.setContentHuggingPriority(.defaultLow, for: .vertical)
+        contentView.weatherGroupLabel.font = .preferredFont(forTextStyle: .largeTitle)
+        contentView.weatherDescriptionLabel.font = .preferredFont(forTextStyle: .largeTitle)
         
         let listInfo = weatherForecastInfo
         let date: Date = Date(timeIntervalSince1970: listInfo.dt)
         navigationItem.title = DateFormatter.convertToKorean(by: date)
-        
-        let iconImageView: UIImageView = UIImageView()
-        let weatherGroupLabel: UILabel = UILabel()
-        let weatherDescriptionLabel: UILabel = UILabel()
-        let temperatureLabel: UILabel = UILabel()
-        let feelsLikeLabel: UILabel = UILabel()
-        let maximumTemperatureLable: UILabel = UILabel()
-        let minimumTemperatureLable: UILabel = UILabel()
-        let popLabel: UILabel = UILabel()
-        let humidityLabel: UILabel = UILabel()
-        let sunriseTimeLabel: UILabel = UILabel()
-        let sunsetTimeLabel: UILabel = UILabel()
-        let spacingView: UIView = UIView()
-        spacingView.backgroundColor = .clear
-        spacingView.setContentHuggingPriority(.defaultLow, for: .vertical)
-        
-        let mainStackView: UIStackView = .init(arrangedSubviews: [
-            iconImageView,
-            weatherGroupLabel,
-            weatherDescriptionLabel,
-            temperatureLabel,
-            feelsLikeLabel,
-            maximumTemperatureLable,
-            minimumTemperatureLable,
-            popLabel,
-            humidityLabel,
-            sunriseTimeLabel,
-            sunsetTimeLabel,
-            spacingView
-        ])
-                
-        mainStackView.arrangedSubviews.forEach { subview in
-            guard let subview: UILabel = subview as? UILabel else { return }
-            subview.textColor = .black
-            subview.backgroundColor = .clear
-            subview.numberOfLines = 1
-            subview.textAlignment = .center
-            subview.font = .preferredFont(forTextStyle: .body)
-        }
-        
-        weatherGroupLabel.font = .preferredFont(forTextStyle: .largeTitle)
-        weatherDescriptionLabel.font = .preferredFont(forTextStyle: .largeTitle)
-        
-        mainStackView.axis = .vertical
-        mainStackView.alignment = .center
-        mainStackView.spacing = 8
-        view.addSubview(mainStackView)
-        mainStackView.translatesAutoresizingMaskIntoConstraints = false
-        
-        let safeArea: UILayoutGuide = view.safeAreaLayoutGuide
-        NSLayoutConstraint.activate([
-            mainStackView.topAnchor.constraint(equalTo: safeArea.topAnchor),
-            mainStackView.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor),
-            mainStackView.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor,
-                                                   constant: 16),
-            mainStackView.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor,
-                                                   constant: -16),
-            iconImageView.widthAnchor.constraint(equalTo: iconImageView.heightAnchor),
-            iconImageView.widthAnchor.constraint(equalTo: safeArea.widthAnchor,
-                                                 multiplier: 0.3)
-        ])
-        
-        weatherGroupLabel.text = listInfo.weather.main
-        weatherDescriptionLabel.text = listInfo.weather.description
-        temperatureLabel.text = "현재 기온 : \(listInfo.main.temp)\(tempUnit.symbol)"
-        feelsLikeLabel.text = "체감 기온 : \(listInfo.main.feelsLike)\(tempUnit.symbol)"
-        maximumTemperatureLable.text = "최고 기온 : \(listInfo.main.tempMax)\(tempUnit.symbol)"
-        minimumTemperatureLable.text = "최저 기온 : \(listInfo.main.tempMin)\(tempUnit.symbol)"
-        popLabel.text = "강수 확률 : \(listInfo.main.pop * 100)%"
-        humidityLabel.text = "습도 : \(listInfo.main.humidity)%"
-        
+
+        contentView.weatherGroupLabel.text = listInfo.weather.main
+        contentView.weatherDescriptionLabel.text = listInfo.weather.description
+        contentView.temperatureLabel.text = "현재 기온 : \(listInfo.main.temp)\(tempUnit.symbol)"
+        contentView.feelsLikeLabel.text = "체감 기온 : \(listInfo.main.feelsLike)\(tempUnit.symbol)"
+        contentView.maximumTemperatureLable.text = "최고 기온 : \(listInfo.main.tempMax)\(tempUnit.symbol)"
+        contentView.minimumTemperatureLable.text = "최저 기온 : \(listInfo.main.tempMin)\(tempUnit.symbol)"
+        contentView.popLabel.text = "강수 확률 : \(listInfo.main.pop * 100)%"
+        contentView.humidityLabel.text = "습도 : \(listInfo.main.humidity)%"
         
         let sunriseDate = Date(timeIntervalSince1970: cityInfo.sunrise)
         let sunsetDate = Date(timeIntervalSince1970: cityInfo.sunset)
         
-        sunriseTimeLabel.text = "일출 : \(DateFormatter.convertToCityTime(by: sunriseDate))"
-        sunsetTimeLabel.text = "일몰 : \(DateFormatter.convertToCityTime(by: sunsetDate))"
+        contentView.sunriseTimeLabel.text = "일출 : \(DateFormatter.convertToCityTime(by: sunriseDate))"
+        contentView.sunsetTimeLabel.text = "일몰 : \(DateFormatter.convertToCityTime(by: sunsetDate))"
         
         let iconName: String = listInfo.weather.icon
         
         imageService.fetchImage(iconName: iconName) { image in
-            DispatchQueue.main.async {
-                iconImageView.image = image
+            DispatchQueue.main.async { [weak self] in
+                guard let self = self else { return }
+                contentView.iconImageView.image = image
             }
         }
     }
